Struct rusty_gaym_engine::matrix::Matrix

pub struct Matrix { /* private fields */ }

Implementations

impl Matrix

pub const fn empty() -> Matrix

Returns matrix with zero columns and zero rows.

pub fn new(
rows_count: usize,
cols_count: usize,
elements: Vec<f32>
) -> Option<Matrix>

Return matrix with specified size and elements.

Examples

let m = Matrix::new(3, 3, vec![
    1.0, 2.0, 3.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0,
]);
assert!(m.is_some());

let m = Matrix::new(3, 3, vec![
    1.0, 2.0, 3.0, 4.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0, 0.0, 10.0,
]);
assert!(m.is_none());

pub fn from_rows(rows: Vec<Vector>) -> Option<Matrix>

Returns matrix with specified rows.

If rows is empty returns empty matrix.

Each row in rows is a Vector, all vectors must be the same size, otherwise function returns None.

pub fn from_cols(cols: Vec<Vector>) -> Option<Matrix>

Returns matrix with specified columns.

If cols is empty returns empty matrix.

Each column in cols is a Vector, all vectors must be the same size, otherwise function returns None.

pub fn from_row(row: &Vector) -> Matrix

Returns matrix made of one row.

pub fn from_col(col: &Vector) -> Matrix

Returns matrix made of one column.

pub fn from_rule<F>(rows_count: usize, cols_count: usize, f: F) -> Matrixwhere
F: Fn(usize, usize) -> f32,

Creates and returns new Matrix instance with the specified number of rows and columns, using the provided closure f to initialize each element in the matrix.

Arguments

rows_count - The number of rows in the matrix

cols_count - The number of columns in the matrix

f - A closure that takes in two usize variables representing the row and column indices of an element, and returns an f32 value to initialize that element.

If rows_count or cols_count is 0 empty matrix will be returned.

Examples

let matrix = Matrix::from_rule(3, 4, |i, j| (i + j) as f32);
 
assert_eq!(matrix, Matrix::from_rows(vec![
    Vector::new(vec![0.0, 1.0, 2.0, 3.0]),
    Vector::new(vec![1.0, 2.0, 3.0, 4.0]),
    Vector::new(vec![2.0, 3.0, 4.0, 5.0])
]).unwrap());

pub fn identity(size: usize) -> Matrix

Returns an identity matrix of the specified size. The identity matrix is a square matrix with ones on the diagonal and zeroes elsewhere.

Arguments

size - The size of the square matrix. If size is zero, empty matrix will be returned.

Example

let identity = Matrix::identity(3);
assert_eq!(identity, Matrix::new(3, 3, vec![1.0, 0.0, 0.0,
                                            0.0, 1.0, 0.0,
                                            0.0, 0.0, 1.0]).unwrap());

pub fn zeroes(rows_count: usize, cols_count: usize) -> Matrix

pub fn gram_matrix(bases: Vec<Vector>) -> Matrix

pub fn adjoint_matrix(&self) -> Option<Matrix>

pub fn rotation_matrix2d(radians: f32) -> Matrix

pub fn rotation_matrix3d(x_radians: f32, y_radians: f32, z_radians: f32) -> Matrix

pub fn get_minor(
&self,
rows_for_exclusion: Vec<usize>,
cols_for_exclusion: Vec<usize>
) -> Matrix

pub fn rows_count(&self) -> usize

pub fn cols_count(&self) -> usize

pub fn get_row(&self, index: usize) -> Option<Vector>

pub fn get_col(&self, index: usize) -> Option<Vector>

pub fn get(&self, row_index: usize, col_index: usize) -> Option<f32>

pub fn determinant(&self) -> Option<f32>

Returns determinant of a square matrix. If matrix is non-square return None.

Algorithm complexity is O(n^3) where n is the order of matrix.

Examples

let m = Matrix::new(3, 3, vec![
    1.0, 2.0, 3.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0,
]).unwrap();

let det = m.determinant().unwrap();
 
assert!(det.abs() < 1e-7);

pub fn inverse(&self) -> Option<Matrix>

pub fn transposed(&self) -> Matrix

pub fn approximately_equal(&self, other: &Matrix, epsilon: f32) -> bool

pub fn set_row(&mut self, row_index: usize, new_row: Vector) -> Result<(), ()>

pub fn set_col(&mut self, col_index: usize, new_col: Vector) -> Result<(), ()>

pub fn set(
&mut self,
row_index: usize,
col_index: usize,
value: f32
) -> Result<(), ()>

pub fn swap_cols(&mut self, a_index: usize, b_index: usize) -> Result<(), ()>

pub fn swap_rows(&mut self, a_index: usize, b_index: usize) -> Result<(), ()>

pub fn multiply(&self, other: &Matrix) -> Option<Matrix>

pub fn multiply_by_vector(&self, other: &Vector) -> Option<Vector>