Struct rusty_gaym_engine::vector::Vector

pub struct Vector {
    pub coordinates: Vec<f32>,
}

Fields
coordinates: Vec<f32>
Implementations

impl Vector

pub const fn new(coordinates: Vec<f32>) -> Vector

pub fn zero3() -> Vector

pub fn one3() -> Vector

pub fn from_xyz(x: f32, y: f32, z: f32) -> Vector

pub fn rotate2d(&self, radians: f32) -> Option<Vector>

Rotates 2d vector counterclockwise on angle expressed in radians. Returns rotated 2d vector. If provided vector is not 2d, returns None.

pub fn rotate3d(
&self,
x_radians: f32,
y_radians: f32,
z_radians: f32
) -> Option<Vector>

Rotates a vector along each of the three axes by the given angles for each axis. Returns rotated 3d vector. If provided vector is not 3d, returns None.

pub fn dim(&self) -> usize

Returns number of components in vector.

pub fn get(&self, index: usize) -> Option<f32>

Access vector component.

pub fn square_distance(&self, coordinates: &Vector) -> f32

Returns the square of Eucledean distance between two vectors.

pub fn distance(&self, b: &Vector) -> f32

Returns the Euclidean distance between two vectors. This method first calculates the square of the distance between the two vectors, and then takes the square root of the result. If you want to get square distance use square_distance method instead.

pub fn square_magnitude(&self) -> f32

Returns the square of Euclidean magnitude of vector.

pub fn magnitude(&self) -> f32

Returns Euclidean magnitude of vector. This method first calculates the square magnitude, and then takes the square root of result. If you want to get square magnitude use square_magnitude method instead.

pub fn dot_product(&self, other: &Vector) -> f32

Computes the dot product of this vector with another vector.

The dot product of two vectors is defined as the sum of the products of their corresponding coordinates.

Examples

let v1 = Vector::new(vec![1.0, 2.0, 3.0]);
let v2 = Vector::new(vec![4.0, 5.0, 6.0]);
assert_eq!(v1.dot_product(&v2), 32.0);

pub fn cross_product(&self, other: &Vector) -> Option<Vector>

Returns cross product of 3d vector and other 3d vector. If vectors are not 3d returns None.

pub fn approximately_equal(&self, other: &Vector, epsilon: f32) -> bool

Vectors are approximately equal if the absolute difference between corresponding vector elements are less than epsilon.

pub fn set(&mut self, index: usize, value: f32) -> Result<(), ()>